+++ src/main/java/com/ffxivcensus/gatherer/GathererController.java
@@ -347,7 +347,7 @@ public class GathererController {
                 sbSQL.append("soundtrack BIT,saweternalbond BIT,sightseeing BIT,comm50 BIT,moogleplush BIT,");
                 sbSQL.append("topazcarubuncleplush BIT,emeraldcarbuncleplush BIT,");
                 sbSQL.append("hildibrand BIT, dideternalbond BIT, arrcollector BIT,");
-                sbSQL.append("kobold BIT, sahagin BIT, amaljaa BIT, sylph BIT, moogle BIT, vanuvanu BIT, ");
+                sbSQL.append("kobold BIT, sahagin BIT, amaljaa BIT, sylph BIT, moogle BIT, vanuvanu BIT, vath BIT, ");
                 sbSQL.append("arr_25_complete BIT,hw_complete BIT, hw_31_complete BIT, hw_33_complete BIT, legacy_player BIT");
             }
             if (this.storeMounts) {
@@ -509,12 +509,12 @@ public class GathererController {
                                 + player.getBitHasEmeraldCarbunclePlush() + ",");
 
                 sbFields.append("hildibrand, ps4collectors, dideternalbond, arrcollector, kobold, sahagin, amaljaa, "
-                                + "sylph, moogle, vanuvanu, hw_complete, hw_31_complete, hw_33_complete, legacy_player");
+                                + "sylph, moogle, vanuvanu, vath, hw_complete, hw_31_complete, hw_33_complete, legacy_player");
                 sbValues.append(player.getBitHasCompletedHildibrand() + "," + player.getBitHasPS4Collectors() + ","
                         + player.getBitHasEternalBond() + "," + player.getBitHasARRCollectors() + ","
-                        + player.getBitHasKobold() + "," + player.getBitHasSahagin() + "," + player.getBitHasAmaljaa()
-                        + "," + player.getBitHasSylph() + "," + player.getBitHasMoogle() + "," + player.getBitHasVanuVanu()
-                        + "," + player.getBitHasCompletedHW() + ","
+                        + player.getBitHasKobold() + "," + player.getBitHasSahagin() + "," + player.getBitHasAmaljaa() + ","
+                        + player.getBitHasSylph() + "," + player.getBitHasMoogle() + "," + player.getBitHasVanuVanu() + ","
+                        + player.getBitHasVath() + "," + player.getBitHasCompletedHW() + ","
                         + player.getBitHasCompleted3pt1() + "," + player.getBitHasCompleted3pt3() + ","
                         + player.getBitIsLegacyPlayer());